+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-29T09:12:35.937Z",
+    "2026-06-29T13:49:49.299Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {